% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.parsuj_skladnie_mcda}
\alias{.parsuj_skladnie_mcda}
\title{Wewnętrzny parser składni MCDA}
\usage{
.parsuj_skladnie_mcda(skladnia)
}
\description{
Zamienia tekst typu "Kryterium =~ zm1 + zm2; Inne =~ zm3" na listę mapowania.
}
\keyword{internal}
