% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.rozmyj_wektor}
\alias{.rozmyj_wektor}
\title{Wewnętrzny fuzzifier TFN}
\usage{
.rozmyj_wektor(wektor)
}
\description{
Zamienia liczbę x na TFN (l,m,u) = (x-1, x, x+1) z ograniczeniem do \verb{1..9}.
Zera (braki) zostają zerami.
}
\keyword{internal}
