% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.skaluj_do_saaty}
\alias{.skaluj_do_saaty}
\title{Wewnętrzny skaler do skali Saaty'ego (1-9)}
\usage{
.skaluj_do_saaty(wektor)
}
\description{
Skaluje dowolny wektor dodatni do przedziału 1..9 (z obsługą NA).
}
\keyword{internal}
