% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwm_solver.R
\name{oblicz_wagi_bwm}
\alias{oblicz_wagi_bwm}
\title{Obliczanie wag kryteriów metodą BWM (Best–Worst Method)}
\usage{
oblicz_wagi_bwm(kryteria, b_to_o, o_to_w)
}
\arguments{
\item{kryteria}{wektor nazw kryteriów (np. c("Reach","Engagement","Cost","Conversion"))}

\item{b_to_o}{wektor długości n: porównania najlepszego kryterium do pozostałych (skala 1..9)}

\item{o_to_w}{wektor długości n: porównania pozostałych do najgorszego (skala 1..9)}
}
\value{
lista: wagi_kryteriow, epsilon, wskaznik_spojnosci, kryteria
}
\description{
Implementacja BWM jako zadania programowania liniowego.
Użytkownik podaje wektor preferencji "Best-to-Others" i "Others-to-Worst".
}
