% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{plot.rozmyty_topsis_wynik}
\alias{plot.rozmyty_topsis_wynik}
\title{Mapa TOPSIS}
\usage{
\method{plot}{rozmyty_topsis_wynik}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy \code{rozmyty_topsis_wynik}}

\item{...}{Ignorowane}
}
\description{
Mapa TOPSIS
}
