% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{plot.rozmyty_waspas_wynik}
\alias{plot.rozmyty_waspas_wynik}
\title{Mapa WASPAS}
\usage{
\method{plot}{rozmyty_waspas_wynik}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy \code{rozmyty_waspas_wynik}}

\item{...}{Ignorowane}
}
\description{
Mapa WASPAS
}
