% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ranking.R
\name{rozmyty_meta_ranking}
\alias{rozmyty_meta_ranking}
\title{Rozmyty meta-ranking}
\usage{
rozmyty_meta_ranking(
  macierz_decyzyjna,
  typy_kryteriow,
  wagi = NULL,
  bwm_kryteria = NULL,
  bwm_najlepsze = NULL,
  bwm_najgorsze = NULL,
  lambda = 0.5,
  v = 0.5
)
}
\arguments{
\item{macierz_decyzyjna}{Rozmyta macierz TFN (m x 3n)}

\item{typy_kryteriow}{Wektor "min"/"max" (długość n)}

\item{wagi}{(opcjonalnie) wagi ręczne (długość n lub 3n)}

\item{bwm_kryteria}{(opcjonalnie) nazwy kryteriów}

\item{bwm_najlepsze}{(opcjonalnie) Best-to-Others}

\item{bwm_najgorsze}{(opcjonalnie) Others-to-Worst}

\item{lambda}{WASPAS lambda}

\item{v}{VIKOR v}
}
\value{
lista: porownanie (data.frame) i korelacje (macierz Spearmana)
}
\description{
Rozmyty meta-ranking
}
