% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_mcda.R
\name{rozmyty_topsis}
\alias{rozmyty_topsis}
\title{Rozmyta metoda TOPSIS}
\usage{
rozmyty_topsis(
  macierz_decyzyjna,
  typy_kryteriow,
  wagi = NULL,
  bwm_kryteria = NULL,
  bwm_najlepsze = NULL,
  bwm_najgorsze = NULL
)
}
\arguments{
\item{macierz_decyzyjna}{Rozmyta macierz TFN (m x 3n).}

\item{typy_kryteriow}{Wektor "min"/"max" dla każdego kryterium (długość n).}

\item{wagi}{(opcjonalnie) Wagi ręczne (długość n lub 3n).}

\item{bwm_kryteria}{(opcjonalnie) Nazwy kryteriów dla BWM.}

\item{bwm_najlepsze}{(opcjonalnie) Wektor Best-to-Others (1-9).}

\item{bwm_najgorsze}{(opcjonalnie) Wektor Others-to-Worst (1-9).}
}
\value{
Obiekt klasy \code{rozmyty_topsis_wynik}.
}
\description{
Implementacja Fuzzy TOPSIS: normalizacja, ważenie, rozwiązanie idealne i anty-idealne,
odległości i współczynnik bliskości (CC), ranking malejąco po CC.
}
