% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyte_mcda.R
\name{rozmyty_waspas}
\alias{rozmyty_waspas}
\title{Rozmyta metoda WASPAS}
\usage{
rozmyty_waspas(
  macierz_decyzyjna,
  typy_kryteriow,
  lambda = 0.5,
  wagi = NULL,
  bwm_kryteria = NULL,
  bwm_najlepsze = NULL,
  bwm_najgorsze = NULL
)
}
\arguments{
\item{macierz_decyzyjna}{Rozmyta macierz TFN (m x 3n).}

\item{typy_kryteriow}{Wektor "min"/"max" dla każdego kryterium (długość n).}

\item{lambda}{Udział WSM vs WPM (domyślnie 0.5).}

\item{wagi}{(opcjonalnie) Wagi ręczne (długość n lub 3n).}

\item{bwm_kryteria}{(opcjonalnie) Nazwy kryteriów dla BWM.}

\item{bwm_najlepsze}{(opcjonalnie) Wektor Best-to-Others (1-9).}

\item{bwm_najgorsze}{(opcjonalnie) Wektor Others-to-Worst (1-9).}
}
\value{
Obiekt klasy \code{rozmyty_waspas_wynik}.
}
\description{
Łączy WSM (suma ważona) i WPM (iloczyn ważony).
Ranking malejąco po wyniku Q.
}
